Metaanalysis, + gwas_qc + folders.config
- Metaanalysis scripts adapted
- Create new positive-controls.txt template
- Few minor changes in some gwas-qc scripts
- Adapt 10_GWAS_QC_multistudies and the corresponding R script

+++ metaanalysis/02-locate-input-files.sh
@@ -1,19 +1,29 @@
-INPUTFILE=input-file-list.txt
-SCRIPTS=/storage/scripts/ckdgenR5/metaanalysis
-LOGFILE=01-locate-input-files.log
-OUTPUTFILE=input-files-with-path.txt
+source ../gwas_qc/folders.config.sh
+
 Phenotype=$1
 
+if [ "$Phenotype" == "" ]
+then
+	echo "Please, give a phenotype name as first argument."
+	exit 3
+fi
+
+
+MA_P_DIR=${PREFIX}/metaanalysis/$Phenotype
+
+
+LOGFILE=${MA_P_DIR}/01-locate-input-files.log
+OUTPUTFILE=${MA_P_DIR}/input-files-with-path.txt
 
 rm -f $LOGFILE
 rm -f $OUTPUTFILE
 
-cut -f 1 $INPUTFILE > input-file-list-col1.txt
+cut -f 1 $INPUT_MA_FILE > ${MA_P_DIR}/input-file-list-col1.txt
 
-bash $SCRIPTS/find-paths.sh input-file-list-col1.txt $OUTPUTFILE $1| tee -a $LOGFILE
+bash $SCRIPTS_MA/find-paths.sh ${MA_P_DIR}/input-file-list-col1.txt $OUTPUTFILE $1| tee -a $LOGFILE
 
 #rm -f input-file-list-col1.txt
 
 echo "TOTAL FILE COUNT (without/with paths)" | tee -a $LOGFILE
 
-wc -l $INPUTFILE $OUTPUTFILE | tee -a $LOGFILE
+wc -l $INPUT_MA_FILE $OUTPUTFILE | tee -a $LOGFILE
